scripts/requirements.txt
numpy
Pillow==9.5.0
moviepy==1.0.3